페이지네이션 구현, 해시태그 제대로 뜨도록 구현, 로그아웃 전 사용자 의사체크 구현

+++ front/public/js/logout.js
@@ -1,6 +1,10 @@
 const logoutBtn = document.querySelector(".logout-btn");
 
 logoutBtn.addEventListener("click", async() => {
+    
+    const logoutConfirm = confirm("로그아웃 하시겠습니까?");
+    if(!logoutConfirm) {return};
+
     try {
         const { data } = await axios.delete("/auth/logout");
         window.location.href = data.message
